improve slot update
